+++ webauthn-server-jvm-crypto/src/test/kotlin/dev/webauthn/server/crypto/AndroidSafetyNetAttestationStatementVerifierTest.kt
@@ -15,6 +15,8 @@ import dev.webauthn.model.RegistrationResponse
 import dev.webauthn.model.RpId
 import dev.webauthn.model.UserHandle
 import dev.webauthn.model.ValidationResult
+import dev.webauthn.model.toNotBlankStringOrThrow
+import dev.webauthn.model.toNotEmptyListOrThrow
 import java.security.KeyPairGenerator
 import java.security.MessageDigest
 import java.security.Signature
@@ -238,10 +240,10 @@ class AndroidSafetyNetAttestationStatementVerifierTest {
     ): RegistrationValidationInput {
         return RegistrationValidationInput(
             options = PublicKeyCredentialCreationOptions(
-                rp = PublicKeyCredentialRpEntity(id = RpId.parseOrThrow("example.com"), name = "Example"),
-                user = PublicKeyCredentialUserEntity(id = UserHandle.fromBytes(ByteArray(16){0}), name = "alice", displayName = "Alice"),
+                rp = PublicKeyCredentialRpEntity(id = RpId.parseOrThrow("example.com"), name = "Example".toNotBlankStringOrThrow("rp.name")),
+                user = PublicKeyCredentialUserEntity(id = UserHandle.fromBytes(ByteArray(16){0}), name = "alice".toNotBlankStringOrThrow("user.name"), displayName = "Alice".toNotBlankStringOrThrow("user.displayName")),
                 challenge = Challenge.fromBytes(ByteArray(16){1}),
-                pubKeyCredParams = emptyList(),
+                pubKeyCredParams = listOf(dev.webauthn.model.PublicKeyCredentialParameters(dev.webauthn.model.PublicKeyCredentialType.PUBLIC_KEY, -7)).toNotEmptyListOrThrow("pubKeyCredParams"),
             ),
             response = RegistrationResponse(
                 credentialId = credentialId,